working: tablet media query header layout finished
I need to work on the other sections in order to
get the screen to line up properly. I was taking the
longest time trying to figure out how to inline
position the stock image in the header in the top
left corner where it looks like its overflow with the
position css property, but that was wrong. I
needed to use the transform property that I totally
forgot. I managed to use transform to correctly
scale, and translate the images and visual elements
of the header.