fix: overflow was mishandled with overflow-x: hidden
So with the other layouts, the diagonal shape
was fine and overflow-x: hidden worked, but
then the mobile was allowed to scroll offscreen
so I set it to instead be overflow-x: clip
so that the diagonal shape can still render
but the user cannot scroll past a certain point.
overflow-x: hidden fully prevented the diagonal
shape from even rendering at all if it overflowed,
so I added a <div> around the shape to scope it
to be positioned within the <div> wrapper.